+++ backend/src/main/java/org/donggle/backend/application/service/auth/MemberService.java
@@ -1,14 +1,10 @@
-package org.donggle.backend.application.service;
+package org.donggle.backend.application.service.auth;
 
 import lombok.RequiredArgsConstructor;
 import org.donggle.backend.application.repository.CategoryRepository;
 import org.donggle.backend.application.repository.MemberCredentialsRepository;
 import org.donggle.backend.application.repository.MemberRepository;
 import org.donggle.backend.application.repository.TokenRepository;
-import org.donggle.backend.domain.oauth.LoginClients;
-import org.donggle.backend.domain.oauth.SocialType;
-import org.donggle.backend.infrastructure.oauth.kakao.dto.response.UserInfo;
-import org.donggle.backend.application.service.request.OAuthAccessTokenRequest;
 import org.donggle.backend.domain.auth.JwtTokenProvider;
 import org.donggle.backend.domain.auth.RefreshToken;
 import org.donggle.backend.domain.category.Category;
@@ -16,6 +12,7 @@ import org.donggle.backend.domain.member.Member;
 import org.donggle.backend.domain.member.MemberCredentials;
 import org.donggle.backend.exception.business.DuplicatedMemberException;
 import org.donggle.backend.exception.notfound.MemberNotFoundException;
+import org.donggle.backend.infrastructure.oauth.kakao.dto.response.UserInfo;
 import org.donggle.backend.ui.response.TokenResponse;
 import org.springframework.dao.DuplicateKeyException;
 import org.springframework.stereotype.Service;
@@ -24,24 +21,14 @@ import org.springframework.transaction.annotation.Transactional;
 @Service
 @Transactional
 @RequiredArgsConstructor
-public class AuthService {
-    private final LoginClients oauthClients;
+public class MemberService {
     private final MemberRepository memberRepository;
     private final JwtTokenProvider jwtTokenProvider;
     private final TokenRepository tokenRepository;
     private final MemberCredentialsRepository memberCredentialsRepository;
     private final CategoryRepository categoryRepository;
 
-    @Transactional(readOnly = true)
-    public String createAuthorizeRedirectUri(final String socialType, final String redirect_uri) {
-        return oauthClients.redirectUri(SocialType.from(socialType), redirect_uri);
-    }
-
-    public TokenResponse login(final String socialType, final OAuthAccessTokenRequest request) {
-        final UserInfo userInfo = oauthClients.findUserInfo(
-                SocialType.from(socialType),
-                request.code(),
-                request.redirect_uri());
+    public TokenResponse login(final UserInfo userInfo) {
         final Member loginMember = memberRepository.findBySocialId(userInfo.socialId())
                 .orElseGet(() -> initializeMember(userInfo));
         return createTokens(loginMember);
@@ -88,4 +75,5 @@ public class AuthService {
                         () -> tokenRepository.save(new RefreshToken(refreshToken, member))
                 );
     }
+
 }
