+++ backend/src/main/java/org/donggle/backend/config/VendorConfiguration.java
@@ -2,8 +2,8 @@ package org.donggle.backend.config;
 
 import org.donggle.backend.application.client.BlogClient;
 import org.donggle.backend.application.client.LoginClient;
+import org.donggle.backend.application.service.auth.LoginClients;
 import org.donggle.backend.domain.blog.BlogClients;
-import org.donggle.backend.domain.oauth.LoginClients;
 import org.donggle.backend.infrastructure.client.notion.NotionApiClient;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
