+++ backend/src/main/java/org/donggle/backend/ui/AuthController.java
@@ -1,6 +1,6 @@
 package org.donggle.backend.ui;
 
-import org.donggle.backend.application.service.AuthService;
+import org.donggle.backend.application.service.auth.AuthService;
 import org.donggle.backend.application.service.request.OAuthAccessTokenRequest;
 import org.donggle.backend.ui.common.AuthenticationPrincipal;
 import org.donggle.backend.ui.response.AccessTokenResponse;
